+++ 3. 클린 코드/클완스 TIL/14장. Args 코드/ArgumentMarshaler.java
@@ -1,0 +1,7 @@
+import java.util.Iterator;
+
+public interface ArgumentMarshaler {
+    public abstract void set(Iterator<String> currentArgument) throws ArgsException;
+
+    public abstract Object get();
+}
