+++ 3. 클린 코드/클완스 TIL/14장. Args 코드/BooleanArgumentMarshaler.java
@@ -1,0 +1,13 @@
+import java.util.Iterator;
+
+public class BooleanArgumentMarshaler implements ArgumentMarshaler {
+    private boolean booleanValue = false;
+
+    public void set(Iterator<String> currentArgument) throws ArgsException {
+        booleanValue = true;
+    }
+
+    public Object get() {
+        return booleanValue;
+    }
+}
