+++ 3. 클린 코드/클완스 TIL/14장. Args 코드/DoubleArgumentMarshaler.java
@@ -1,0 +1,22 @@
+import java.util.Iterator;
+import java.util.NoSuchElementException;
+
+public class DoubleArgumentMarshaler implements ArgumentMarshaler {
+    private double doubleValue = 0;
+
+    public void set(Iterator<String> currentArgument) throws ArgsException {
+        String parameter = null;
+        try {
+            parameter = currentArgument.next();
+            doubleValue = Double.parseDouble(parameter);
+        } catch (NoSuchElementException e) {
+            throw new ArgsException(ArgsException.ErrorCode.MISSING_DOUBLE);
+        } catch (NumberFormatException e) {
+            throw new ArgsException(ArgsException.ErrorCode.INVALID_DOUBLE, parameter);
+        }
+    }
+
+    public Object get() {
+        return doubleValue;
+    }
+}
