+++ 3. 클린 코드/클완스 TIL/14장. Args 코드/IntegerArgumentMarshaler.java
@@ -1,0 +1,22 @@
+import java.util.Iterator;
+import java.util.NoSuchElementException;
+
+public class IntegerArgumentMarshaler implements ArgumentMarshaler {
+    private int intValue = 0;
+
+    public void set(Iterator<String> currentArgument) throws ArgsException {
+        String parameter = null;
+        try {
+            parameter = currentArgument.next();
+            intValue = Integer.parseInt(parameter);
+        } catch (NoSuchElementException e) {
+            throw new ArgsException(ArgsException.ErrorCode.MISSING_INTEGER);
+        } catch (NumberFormatException e) {
+            throw new ArgsException(ArgsException.ErrorCode.INVALID_INTEGER, parameter);
+        }
+    }
+
+    public Object get() {
+        return intValue;
+    }
+}
