+++ 3. 클린 코드/클완스 TIL/14장. Args 코드/StringArgumentMarshaler.java
@@ -1,0 +1,19 @@
+import java.util.Iterator;
+import java.util.NoSuchElementException;
+
+public class StringArgumentMarshaler implements ArgumentMarshaler {
+    private String stringValue = "";
+
+    public void set(Iterator<String> currentArgument) throws ArgsException {
+        try {
+            stringValue = currentArgument.next();
+        } catch (NoSuchElementException e) {
+            throw new ArgsException(ArgsException.ErrorCode.MISSING_STRING);
+        }
+
+    }
+
+    public Object get() {
+        return stringValue;
+    }
+}
