--- 3. 클린 코드/클완스 TIL/refactor-example.java
@@ -1,286 +1,0 @@
-package com.objectmentor.utilities.getopts;
-
-import java.lang.reflect.Parameter;
-import java.text.ParseException;
-import java.util.*;
-
-public class Args {
-    private String schema;
-    private Set<Character> unexpectedArguments = new TreeSet<Character>();
-    private Map<Character, ArgumentMarshaler> marshalers = new HashMap<Character, ArgumentMarshaler>();
-    private Set<Character> argsFound = new HashSet<Character>();
-    private Iterator<String> currentArgument;
-    private List<String> argsList;
-
-    public Args(String schema, String[] args) throws ArgsException {
-        this.schema = schema;
-        argsList = Arrays.asList(args);
-        parse();
-    }
-
-    private void parse() throws ArgsException {
-        parseSchema();
-        parseArguments();
-    }
-
-    private void parseSchema() throws ArgsException {
-        for (String element : schema.split(",")) {
-            if (element.length() > 0) {
-                String trimmedElement = element.trim();
-                parseSchemaElement(trimmedElement);
-            }
-        }
-    }
-
-    private void parseSchemaElement(String element) throws ArgsException {
-        char elementId = element.charAt(0);
-        String elementTail = element.substring(1);
-        validateSchemaElementId(elementId);
-        if (elementTail.length() == 0)
-            marshalers.put(elementId, new BooleanArgumentMarshaler());
-        else if (elementTail.equals("*"))
-            marshalers.put(elementId, new StringArgumentMarshaler());
-        else if (elementTail.equals("#"))
-            marshalers.put(elementId, new IntegerArgumentMarshaler());
-        else if (elementTail.equals("##"))
-            marshalers.put(elementId, new DoubleArgumentMarshaler());
-        else {
-            throw new ArgsException(ArgsException.ErrorCode.INVALID_FORMAT, elementId, elementTail);
-        }
-    }
-
-    private void validateSchemaElementId(char elementId) throws ArgsException {
-        if (!Character.isLetter(elementId)) {
-            throw new ArgsException(ArgsException.ErrorCode.INVALID_ARGUMENT_NAME, elementId, null);
-        }
-    }
-
-    private void parseArguments() throws ArgsException {
-        for (currentArgument = argsList.iterator(); currentArgument.hasNext();) {
-            String arg = currentArgument.next();
-            parseArgument(arg);
-        }
-    }
-
-    private void parseArgument(String arg) throws ArgsException {
-        if (arg.startsWith("-"))
-            parseElements(arg);
-    }
-
-    private void parseElements(String arg) throws ArgsException {
-        for (int i = 1; i < arg.length(); i++)
-            parseElement(arg.charAt(i));
-    }
-
-    private void parseElement(char argChar) throws ArgsException {
-        if (setArgument(argChar))
-            argsFound.add(argChar);
-        else {
-            throw new ArgsException(ArgsException.ErrorCode.UNEXPECTED_ARGUMENT, argChar, null);
-        }
-    }
-
-    private boolean setArgument(char argChar) throws ArgsException {
-        ArgumentMarshaler m = marshalers.get(argChar);
-        if (m == null)
-            return false;
-        try {
-            m.set(currentArgument);
-            return true;
-        } catch (ArgsException e) {
-            e.setErrorArgumentId(argChar);
-            throw e;
-        }
-    }
-
-    public int cardinality() {
-        return argsFound.size();
-    }
-
-    public String usage() {
-        if (schema.length() > 0)
-            return "-[" + schema + "]";
-        else
-            return "";
-    }
-
-    public boolean getBoolean(char arg) {
-        Args.ArgumentMarshaler am = marshalers.get(arg);
-        boolean b = false;
-        try {
-            b = am != null && (Boolean) am.get();
-        } catch (ClassCastException e) {
-            b = false;
-        }
-        return b;
-    }
-
-    public String getString(char arg) {
-        Args.ArgumentMarshaler am = marshalers.get(arg);
-        try {
-            return am == null ? "" : (String) am.get();
-        } catch (ClassCastException e) {
-            return "";
-        }
-    }
-
-    public int getInt(char arg) {
-        Args.ArgumentMarshaler am = marshalers.get(arg);
-        try {
-            return am == null ? 0 : (Integer) am.get();
-        } catch (ClassCastException e) {
-            return 0;
-        }
-    }
-
-    public double getDouble(char arg) {
-        Args.ArgumentMarshaler am = marshalers.get(arg);
-        try {
-            return am == null ? 0 : (Double) am.get();
-        } catch (ClassCastException e) {
-            return 0.0;
-        }
-    }
-
-    public boolean has(char arg) {
-        return argsFound.contains(arg);
-    }
-
-    private interface ArgumentMarshaler {
-
-        public abstract void set(Iterator<String> currentArgument) throws ArgsException;
-
-        public abstract Object get();
-
-    }
-
-    private class BooleanArgumentMarshaler extends ArgumentMarshaler {
-        private boolean booleanValue = false;
-
-        public void set(Iterator<String> currentArgument) throws ArgsException {
-            booleanValue = true;
-        }
-
-        public Object get() {
-            return booleanValue;
-        }
-    }
-
-    private class StringArgumentMarshaler extends ArgumentMarshaler {
-        private String stringValue = "";
-
-        public void set(Iterator<String> currentArgument) throws ArgsException {
-            try {
-                stringValue = currentArgument.next();
-            } catch (NoSuchElementException e) {
-                throw new ArgsException(ArgsException.ErrorCode.MISSING_STRING);
-            }
-
-        }
-
-        public Object get() {
-            return stringValue;
-        }
-    }
-
-    private class IntegerArgumentMarshaler extends ArgumentMarshaler {
-        private int intValue = 0;
-
-        public void set(Iterator<String> currentArgument) throws ArgsException {
-            String parameter = null;
-            try {
-                parameter = currentArgument.next();
-                intValue = Integer.parseInt(parameter);
-            } catch (NoSuchElementException e) {
-                errorCode = ArgsException.ErrorCode.MISSING_INTEGER;
-                throw new ArgsException();
-            } catch (NumberFormatException e) {
-                throw new ArgsException(ArgsException.ErrorCode.INVALID_INTEGER, errorParameter);
-            }
-        }
-
-        public Object get() {
-            return intValue;
-        }
-    }
-
-    private class DoubleArgumentMarshaler extends ArgumentMarshaler {
-        private double doubleValue = 0;
-
-        public void set(Iterator<String> currentArgument) throws ArgsException {
-            String parameter = null;
-            try {
-                parameter = currentArgument.next();
-                doubleValue = Double.parseDouble(parameter);
-            } catch (NoSuchElementException e) {
-                errorCode = ArgsException.ErrorCode.MISSING_DOUBLE;
-                throw new ArgsException();
-            } catch (NumberFormatException e) {
-                throw new ArgsException(ArgsException.ErrorCode.INVALID_DOUBLE, parameter);
-            }
-        }
-
-        public Object get() {
-            return doubleValue;
-        }
-    }
-}
-
-public class ArgsException extends Exception {
-    private char errorArgumentId = '\0';
-    private String errorParameter = "TILT";
-    private ErrorCode errorCode = ErrorCode.OK;
-
-    public ArgsException() {
-    }
-
-    public ArgsException(String message) {
-        super(message);
-    }
-
-    public ArgsException(ErrorCode errorCode) {
-        this.errorCode = errorCode;
-    }
-
-    public ArgsException(ErrorCode errorCode, String errorParameter) {
-        this.errorCode = errorCode;
-        this.errorParameter = errorParameter;
-    }
-
-    public ArgsException(ErrorCode errorCode, char errorArgumentId, String errorParameter) {
-        this.errorCode = errorCode;
-        this.errorArgumentId = errorArgumentId;
-        this.errorParameter = errorParameter;
-    }
-
-    public void setErrorArgumentId(char errorArgumentId) {
-        this.errorArgumentId = errorArgumentId;
-    }
-
-    public String errorMessage() throws Exception {
-        switch (errorCode) {
-            case UNEXPECTED_ARGUMENT:
-                return String.format("Argument -%c unexpected.", errorArgumentId);
-            case MISSING_STRING:
-                return String.format("Could not find string parameter for -%c.", errorArgumentId);
-            case INVALID_INTEGER:
-                return String.format("Argument -%c expects an integer but was '%s'.", errorArgumentId,
-                        errorParameter);
-            case MISSING_INTEGER:
-                return String.format("Could not find integer parameter for -%c.", errorArgumentId);
-            case INVALID_DOUBLE:
-                return String.format("Argument -%c expects an double but was '%s'.", errorArgumentId, errorParameter);
-            case MISSING_DOUBLE:
-                return String.format("Could not find double parameter for -%c.", errorArgumentId);
-            case OK:
-                throw new Exception("TILT: Should not get here.");
-        }
-        return "";
-    }
-
-    enum ErrorCode {
-        OK, INVALID_FORMAT, INVALID_ARGUMENT_NAME, MISSING_STRING, MISSING_INTEGER, INVALID_INTEGER, MISSING_DOUBLE,
-        INVALID_DOUBLE,
-        UNEXPECTED_ARGUMENT
-    }
-}
